+++ packages/plugin-commons/web/src/Base/toolbars/createAtomicPluginToolbar.jsx
@@ -293,12 +293,11 @@ export default function createAtomicPluginToolbar({
             <BlockSpoilerButton {...commonButtonProps} tooltipText={t('Spoiler_Insert_Tooltip')} />
           );
         case BUTTONS.VIDEO_SETTINGS: {
-          const isCustomVideo = !!this.state.componentData.isCustomVideo;
           const videoSettingsProps = {
             ...defaultButtonProps,
             type: BUTTONS.EXTERNAL_MODAL,
           };
-          return isCustomVideo ? <Button {...videoSettingsProps} /> : null;
+          return <Button {...videoSettingsProps} />;
         }
         case BUTTONS.LINK_PREVIEW: {
           return (