+++ packages/plugin-commons/web/src/index.ts
@@ -31,17 +31,12 @@ export { default as createBaseInsertPluginButton } from './Base/createBaseInsert
 export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
-export { default as FileInput } from './Components/FileInput';
 export { default as Image } from './Components/Image';
 export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
 export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
 export { default as SelectionList } from './Components/SelectionList';
 export { default as SettingsSection } from './Components/SettingsSection';
-export { default as Slider } from './Components/Slider';
-export { default as SliderWithInput } from './Components/SliderWithInput';
 export { Tab, Tabs } from './Components/Tabs';
-export { default as TextInput } from './Components/TextInput';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
-export { default as Loader } from './Components/Loader';
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';