+++ packages/plugin-image/web/src/image-component.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Loader } from 'wix-rich-content-plugin-commons';
-import { MediaItemErrorMsg } from 'wix-rich-content-ui-components';
+import { MediaItemErrorMsg, Loader } from 'wix-rich-content-ui-components';
 import ImageViewer from './image-viewer';
 import { DEFAULTS } from './consts';
 import { sizeClassName, alignmentClassName } from './classNameStrategies';